Update pdus.c mcc uint8_t to uint_16_t

+++ src/pdus.c
@@ -16,7 +16,7 @@ uint8_t pdu_mac_data[16];       /* 124 bits */
 uint8_t pdu_acc_ass[2];
 uint8_t pdu_acc_ass_18[2];
 
-void dm_sync_pdu_schh(uint16_t address, uint8_t rep_mcc, uint16_t rep_mnc){
+void dm_sync_pdu_schh(uint16_t address, uint16_t rep_mcc, uint16_t rep_mnc){
 	struct bitvec bv;
 	memset(&bv, 0, sizeof(bv));
 	bv.data = pdu_sync_SCHH;
@@ -75,7 +75,7 @@ void dm_sync_pres_pdu_schs(uint8_t tn, uint8_t fn, uint8_t frame_countdown, uint
 
 
 
-void sync_pdu(uint16_t cc, uint8_t mn, uint8_t fn, uint8_t tn, uint8_t mcc, uint16_t mnc)
+void sync_pdu(uint16_t cc, uint8_t mn, uint8_t fn, uint8_t tn, uint16_t mcc, uint16_t mnc)
 {
 	struct bitvec bv;
 	memset(&bv, 0, sizeof(bv));
